Refactor, deps, docs improvements, licensing, contrib.
