--- .pdm-python
@@ -1,1 +1,0 @@
-C:/Users/emilh/Desktop/Git/autogen_websocket/.venv/Scripts/python.exe
