+++ CONTRIBUTING.md
@@ -1,0 +1,64 @@
+# Contributing to Agentsflow
+
+First off, thank you for considering contributing to Agentsflow. It's people like you that make Agentsflow such a great tool.
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code.
+
+## How Can I Contribute?
+
+### Reporting Bugs
+
+This section guides you through submitting a bug report for Agentsflow. Following these guidelines helps maintainers and the community understand your report, reproduce the behavior, and find related reports.
+
+- **Use a clear and descriptive title** for the issue to identify the problem.
+- **Describe the exact steps which reproduce the problem** in as many details as possible.
+- **Provide specific examples to demonstrate the steps**. Include links to files or GitHub projects, or copy/pasteable snippets, which you use in those examples.
+
+### Suggesting Enhancements
+
+This section guides you through submitting an enhancement suggestion for Agentsflow, including completely new features and minor improvements to existing functionality.
+
+- **Use a clear and descriptive title** for the issue to identify the suggestion.
+- **Provide a step-by-step description of the suggested enhancement** in as many details as possible.
+- **Describe the current behavior and explain which behavior you expected to see instead** and why.
+
+### Your First Code Contribution
+
+Unsure where to begin contributing to Agentsflow? You can start by looking through these `beginner` and `help-wanted` issues:
+
+- `Beginner issues` - issues which should only require a few lines of code, and a test or two.
+- `Help wanted issues` - issues which should be a bit more involved than `beginner` issues.
+
+### Pull Requests
+
+- Do not include issue numbers in the PR title.
+- Include screenshots in your pull request whenever possible.
+- End all files with a newline.
+- Include appropriate unit tests.
+
+## Styleguides
+
+### Git Commit Messages
+
+- Use the present tense ("Add feature" not "Added feature")
+- Use the imperative mood ("Move cursor to..." not "Moves cursor to...")
+- Limit the first line to 72 characters or less
+- Reference issues and pull requests liberally after the first line
+
+### Python Styleguide
+
+All Python must adhere to [PEP 8](https://www.python.org/dev/peps/pep-0008/).
+
+### JavaScript Styleguide
+
+All JavaScript must adhere to [Airbnb JavaScript Style Guide](https://github.com/airbnb/javascript).
+
+## Get Involved
+
+We tend to congregate on the [autogen Discord server](https://discord.gg/Dav2SEW9UD).
+
+## License
+
+By contributing, you agree that your contributions will be licensed under its GPL-3.0 License.
