--- autogen_client/agents/__init__.py
@@ -1,2 +1,0 @@
-from autogen_client.agents.ws_user_proxy import UserProxyAgent
-from autogen_client.agents.ws_assistant import WebSocketAssistantAgent
