+++ requirements.txt
@@ -1,0 +1,42 @@
+# This file is @generated by PDM.
+# Please do not edit it manually.
+
+aiohttp==3.8.5
+aiosignal==1.3.1
+annotated-types==0.5.0
+anyio==3.7.1
+async-timeout==4.0.3
+attrs==23.1.0
+certifi==2023.7.22
+charset-normalizer==3.3.0
+click==8.1.7
+diskcache==5.6.3
+fastapi==0.103.2
+flaml==2.1.1
+frozenlist==1.4.0
+h11==0.14.0
+httptools==0.6.0
+idna==3.4
+jinja2==3.1.2
+MarkupSafe==2.1.3
+multidict==6.0.4
+NumPy==1.25.2
+openai==0.28.1
+pyautogen==0.1.7
+pydantic==2.4.2
+pydantic-core==2.10.1
+python-dotenv==1.0.0
+pyyaml==6.0.1
+requests==2.31.0
+sniffio==1.3.0
+starlette==0.27.0
+termcolor==2.3.0
+tqdm==4.66.1
+typing-extensions==4.8.0
+urllib3==2.0.6
+uvicorn==0.23.2
+uvloop==0.17.0
+watchfiles==0.20.0
+websockets==11.0.3
+yarl==1.9.2
+
